Move get_names_from_tree
### Description
Will need this to parse a tree after tree building.

#### Issue
[ch126782](https://app.clubhouse.io/genepi/story/126782/workflow-to-pull-the-data-from-aspen-api-endpoint-and-launch-nextstrain-build)

### Test plan
visually.
