+++ src/py/aspen/phylo_tree/identifiers.py
@@ -1,0 +1,11 @@
+from typing import Set
+
+
+def get_names_from_tree(tree) -> Set[str]:
+    results: Set[str] = set()
+    for node in tree:
+        results.add(node["name"])
+        if "children" in node:
+            results.update(get_names_from_tree(node["children"]))
+
+    return results
